[LOG-11300] Fix tests and failures
Tests are completely broken, the run command will always fail because of
not receiving a region in the arguments.

Tests will not catch it since docker ps is grepped and then piped into
wc, which will always succeed, this never caught the failure.

Versioning was broken too. Since we are running node 0.12, const is not
allowed:
https://stackoverflow.com/questions/22603078/syntaxerror-use-of-const-in-strict-mode

Const was being used in some packages due to the `^` in the
package.json, `^` means get the latest version with the same major
number, but within that latest version const was introduced and was
causing failures.